requirements.txt
fastapi
uvicorn
pandas
pillow
python-multipart
onnxruntime
numpy
torch
torchvision
faiss-cpu
git+https://github.com/openai/CLIP.git
